requirements.txt
Django==4.2
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
